Integrate the different metamodels in JarvisCore
This commits contains a large refactoring of the JarvisCore class and the associated Intent processing / Action execution classes.
In particular, the DialogFlowApi now returns RecognizedIntents instead of raw intents, and JarvisCore is constructed from a OrchestrationModel, that is used to enable JarvisModules, load JarvisActions, and generate actions from received RecognizedIntents.
Check the documentation for more information.