Add InputProvider#close() and create InputProvider getter in JarvisCore
The close method allows to properly close the output stream of the InputProvider, and disconnect the associated input stream. 
The InputProvider getter is designed for testing purposes,  and should not be accessed from client code (see documentation). This commit also adds a call to InputProvider#close() in JarvisCore#shutdown() to disconnect the underlying streams when the JarvisCore is shut down.