add first version of the orchestration grammar
This commit provides an IDE to write Orchestration models. The IDE contains advanced features such as automatic linking of intent/module elements, syntax completion of Intent/Action names and parameters, and dynamic verification of the model. 
Saving a .orchestration file produces a corresponding .xmi that can be used as OrchestrationModel input for JarvisCore.