add dependency/plugin version update in the build (fix #81)
The possible dependency and plugin updates are displayed at the end of the build.
Note that there is an issue in the maven plugin that does not allow to set the minimum version to use (https://web.archive.org/web/20150514123930/http://jira.codehaus.org/browse/MVERSIONS-272). The report contains multiple possible plugin updates depending on the maven version that can be chosen.