add support for Action variants in Orchestration model
This allows to specify multiple constructors for the same JarvisAction, and call the right one according to the provided parameters. The implementation at the linker level assumes that there is only one Action with a given name in each module (which is generally the case). If the Action Parameter does not match the provided ones then alternative Action with the same name are searched.