add JarvisMessageAction#getClientSession and merge contexts (fix #128)
JarvisMessageAction subclass should now provide a getClientSession method, that is used to uniquely identify the target of the message to send. This information is used to create the Client session, that is used later to store client-related values. The created session is filled with the content of the JarvisAction session if it is different, in order to pass values from client-independent events.