Add IntentRecognitionProvider interface (refs #140)
This interface is a first step towards the abstraction of intent recognition providers under a common API. The next phase is to remove core dependencies to specific DialogFlow classes, and replace them with the created interface. A factory should also be provided to build concrete IntentRecognitionProviders from a given configuration file.