add GenericChatModule, IntentProvider, and Action (fix #53)
The GenericChatModule reads the provided configuration and initializes a set of sub-modules that interacts with the concrete chatting platforms. This architecture allows to define fully generic orchestration models, that can be deployed in concrete chatting platforms by setting the corresponding properties in the Configuration. Note that this commit does not define a generic Reply action, but an initial Generic postMessage action. The postMessage channel parameter can be specified in the Configuration for each concrete chatting platform.