add support for follow-up intents (fix #47)
This commit defines a new keyword in the language that allows to specify "follow-up" relationship between IntentDefinitions. This new construct is backed by new references in the intent metamodel, and are used in the DialogFlowAPI to create follow-up intents and set their contexts.
Note that the follow-up contexts are not defined at the metamodel/model level, and their creation is the responsability of the DialogFlow API.
Customization of follow-up context lifespan is left for a future release (refs #147)