add lifespan setting and retrieval to JarvisContext (fix #112)
This commit enhances JarvisContext with support for lifespan count setting and retrieval. Setting a context value now requires to specify its lifespan, and JarvisContext#decrementLifespanCounts() can be used to decrease all the lifespan counts of a given context. JarvisContext#merge is also updated to take into account the lifespan values associated to the context.
Note that this commit also changes the variable policy, that are now created in a context with a lifespan of 1, meaning that they are deleted before each user operation. This forces the variable scope to the current event, and avoid to propagate unnecessary variables to the IntentRecognitionProvider.
Context lifespans are decremented when calling IntentProvider#getRecognizedIntent, meaning that variables set by the IntentProviders must be defined AFTER calling this method (see the method documentation for further information)
Additional information on the lifespan management is provided in the JarvisContext documentation.