add JarvisActionResult class (fix #65)
This commit changes the signature of JarvisAction#call, that return a JarvisActionResult. The method is now final, wrapping the raw computation result into JarvisActionResult. JarvisAction now defines an additional compute() method, that is implemented by all the subclasses, and defines the raw computation of the action.
This commit also changes the action execution behavior, and moves the JarvisCore#handleEvent method to the OrchestrationService, which now handles all the EventInstance processing. JarvisAction computation is now performed sequentially instead of the previous Future-based implementation, and all the JarvisActions associated to a given EventInstance are wrapped in a single Future that is executed in a dedicated thread. This should ensure variable consistency when tasks rely on variables defined by other ones.