moved core modules to a dedicated maven project (refs #58)
This commit is a first step towards solving the dynamic import of core modules from the language editor. The new maven project will be used by both the core component and the language editors, that will rely on the pathmap mechanism to allow dynamic loading of core module resources from different locations. This architecture will be also used to dynamically load user-defined modules.