JarvisCore supports dynamic loading of core Modules (fix #58)
This commit complements e73a2f81b22c16ff3948bcc2f0be5705ec270d57 and provides dynamic core module loading in JarvisCore. Note that *all* the core modules are loaded when starting the Jarvis bot. 
Loaded resources are stored in the ResourceSet that is used to load the Orchestration resource, and ensures that the proxies to core modules are well resolved.