JarvisCore supports dynamic loading of custom Modules (fix #158)
This commit changes the Orchestration language and adds an ImportDeclaration object, that defines the import path and an optional alias. If an alias is defined, the references to the model will be stored using a dedicated pathmap, that can be resolved by the JarvisCore component. 
Custom module pathmap resolution is done by inspecting the configuration file. Each alias that is defined in the Orchestration model should be specified in the configuration file following the pattern `JarvisCore#CUSTOM_MODULES_KEY_PREFIX.<module alias> = <module path>`, where `<module path>` is the concrete path to the module model. This architecture makes Orchestration models with alias portables, and allows the JarvisCore component to open them even if they have been moved to a different location.
Note that the JarvisCore component still supports path-based module imports, but does not provide any resolution mechanisms if the path cannot be found.
The JarvisCore component does not provide support for alias models stored in an external jar yet (refs #162).
Note that aliases are ignored when applied to core modules.