add update site for Jarvis DSLs editors (fix #30)
This commit adds the `feature` and `update` projects that are used to build the update site that can be installed in client Eclipse applications. The update site is self-contained, and does not require any preliminary installation of Jarvis components.
The orchestration and module plugins now contain an Activator, that is used to register the required metamodels to the global registry. This is required because the metamodels are built as OSGI bundles, and thus cannot use the Equinox extension registry mechanism.