JarvisMessageAction#compute now retries on IOException (fix #135)
The retry logic is defined in the JarvisMessageAction class, and retries 3 time to send the message, using an increasing waiting time between each attempt.
This behavior should be configurable through the global jarvis configuration (refs #170).
Note that the retry logic is triggered if the underlying computation threw an IOException. This commit adapts the related Slack and Discord actions to throw such exception when an error occurred.
The new implementation of the Discord PostMessage class now waits for the Discord platform response before returning, allowing to detect connection issues.