Rename Module model to Platform (fix #172)
This commit updates the `module` metamodel at the root of the project, as well as the `module` xtext-editor. 
The new `Platform` metamodel define a top-level `Platform` element, replacing the previous `Module`.
The core component is also updated to support these new types, but some refactoring is still needed to remove `module` occurrences from the project.