removed IntentDefinition from Platform (fix #173)
This commit removes the direct references to IntentDefinitions in the Platform metamodel. It is still possible to define EventDefinitions in Platform models through the EventProvider element, that provides a set of Events that can be used as inputs to trigger actions. Putting these EventDefinitions in a dedicated Library doesn't make a lot of sense, since they are strongly coupled to their containing provider.