factorize buildConfiguration methods
The different buildConfiguration methods now call each other to avoid code duplication. For example the JarvisCoreTest#buildConfiguration method relies on the DialogFlowApiTest#buildConfiguration one, to ensure that the DialogFlow credentials are valid.
This commit also adds DialogFlow variables in the test-configuration file (not pushed), such as the default language and the path to the credential file. This should simplify tests on new machines.
DialogFlowApi now attempts to load the provided credential file from the classpath if all the other strategies failed. This allows to easily provide file names in the configuration.