Platform DSL now supports platform extension (refs #176)
This commit add the `extends` and `abstract` keywords to the Platform language, as well as an import mechanism similar to the one in the Execution DSL. 
Platforms can only extends abstract platforms, and the generator takes care of copying the parent actions and providers to the child. This copy behavior is the easiest to implement with the current engine, but we may think of alternative solutions that avoid this kind of duplication.
This commit does not fix the issue: the core module still needs to be updated to support abstract platforms. This will certainly imply to remove the `Generic_Chat` module, which won't be required anymore.