add EventProvider out Context (fix #190)
These contexts are not used for now, but will be employed to provide efficient completion proposals (refs #114)
Note that the core component does not check that context parameters set by the providers are defined in the model, this would add too much restriction on developing new providers, and the current architecture does not allow to do it easily. In the end, this information is only useful when designing the bot through the editor.