add warning log for invalid context fragments (fix #194)
The DialogFlowApi now logs a warning if a context parameter fragment does not correspond to a valid value of its associated mapping entity. DialogFlow checks are performed in a new dedicated class (DialogFlowCheckingUtils) that will gather all the checking methods.
This commit also update the maven-bundle-plugin configuration of the metamodels to export the xcore lib. This fixed the NoClassDefFound at the editor level and allows to call complex xcore operations.