improved intent language keywords
This commit removes unnecessary keywords (e.g. params after outContext) and changes others to improve the reading / writing of intent libraries.
The new version of the language is more verbose, but easier to understand and reuse. Since the language is not designed to be manipulated by experts in intent recognition platform we chose to simplify it and make obvious things that are assumed by the recognition engines.