moved ImportDeclaration in common and reused it in execution
The execution metamodel and language now reuse common::ImportDeclaration, that can also be reused in the other languages.
In addition, the expressions embedded in the execution language (e.g. VariableAccesses) are now reused from the common metamodel. The core component needs to be updated to integrate an interpreter for such expressions.