Add abstract ChatPlatform and support for chat context (fix #212)
This commit adds the RuntimeEventProvider#sendEventInstance method, that should be used by subclasses to forward EventInstances to the core component and trigger actions. This method can be extended by subclasses (e.g. the ChatPlatform) to perform pre-processing checks and ensure that all the variables have been set by the sublcass.