improve editor import resolving performance (fix #186)
The generators now update the registry with the new models once they are produced, this ensures that the registry always contains updated version of the imported resources.
The loading algorithm has been reworked to limit the number of unnecessary file accesses. This should partially solve concurrency issues in the ResourceSet used to load the Resources, and fix unreliable import loading (e.g. imports loaded sometimes and not other times without explanation).