fix context overriding issue
Still not convinced if this is a good idea: right now client.merge(other) updates client with all the variables from other. If a variable is defined in both contexts the one defined in other is used to override the client one.
The main drawback here is that previously set variables are never erased. This may be an issue if other defines less variables that client, because the old client ones can still be accessed through the context() operator.
This should only have an impact on EventProviders, IntentProviders use decrementLifeSpan that should protect context from old variables.