refactor eclipse target platforms
We now have one target platform in eclipse/target-platform that is reused by all the projects. This includes the features and update projects, that were not using it before. This should solve dependency error related to the target platform with specific .m2 configuration (see discussion with @nerlichman)
The created target platform should probably specify the antlr version (refs #236) but it needs to be tested.