simplify provider hierarchy (fix #221)
The RuntimeIntentProvider class does not exist anymore, it has been replaced by the IntentRecognitionHelper that provide a simple interface to extract RecognizedIntents from text.

This means that conceptually now all the providers are subclasses of RuntimeEventProvider, and may or may not use the helper to create intents instead of events.