uniformized RestHandler and providers (fix #219)
This commit removes the JsonWebhookEventProvider, and defines a new RestHandler hierarchy that can be used by WebhookEventProviders to handle Rest requests. 

This refactoring uniformizes the registration of WebhookEventProviders and RestHandlers (both are now registered as handlers), and simplifies the XatkitServer code.

Note that now providers and handlers can return Object, that will be parsed by the server to create the HTTP response body.

WebhookProviders now need to define an URI they are registered to (fix #108). This may imply some refactoring in client code and some reconfiguration in existing webhooks.