Monitoring data folder is now relative to the properties file (fix #243)
This commit adds the FileUtils class, that contains the relative path resolution logic used across Xatkit. FileUtils#getFile will return the File if the provided path is absolute, or try to resolve it against the .properties file containing directory if it is specified. If everything else fails it will resolve relative path from the execution directory.