JsonEventMatcher prints the builder content even if the event is null
This behavior is activated if the `xatkit.event.matcher.print_builder` property is set to `true` in the configuration.
The content of the builder is printed for `null` event to ease the process of creating new events: the builder content can be copy-pasted in the platform editor to add the event to the platform.