add support for Http Get method in XatkitServer (fix #222)
This  commit adds support for all the Http methods supported by http-core. Request parameters are also correctly extracted, and requests with parameters are correctly matched to their base URI.
This commit changes the monitoring API HttpMethod to GET (previously POST)
This commit breaks the public API: it is now required to provide the HttpMethod to register a REST endpoint.