add EmptyContentRestHandler and allow HttpEntity in handler return
This commit allows to define handlers that return HttpEntities (e.g. the content of an HTML response). In this case the creation of the HttpEntity is the responsability of the handler.
This commit also adds a new exception type RestHandlerException, that lets handlers notify the server that an error occurred.