integrate StanfordNLP for basic sentiment analysis and yes/no questions
This commits adds two post-processors that could be used to extract the user's sentiment and whether the input is a yes/no question. The analysis result can be found in the context.
These post-processors require to install StanfordNLP in Xatkit, follow the wiki's instruction to learn more about it.
Fix #275
Fix #276