removed session switch within the same state's body
Some action create their own session (e.g. PostMessage) that match the client they target. When this happens we need to merge the current session in the created one, in order to make sure all the variables are accessible to the action's session.
The removed code was also switching the body's session to the action's one after the computation of the action, in order to make sure setting new variables in the session after the action computation was working fine.
The goal was to support something like this
`session 1`
`SlackPlatform.PostMessage(...) // creates session 2`
`session.put() // adds the value in session 2`
This is a common use case when receiving events and forwarding them to a chat platform. If the body's session is not switched session 2 will not be updated but session 1 will.
While this is pretty handy for the use case above, this is EXTREMELY MISLEADING for  the bot designer, who cannot know at design time if the session has been switched.

With this commit the body's session is NEVER switched. The main restriction it implies is to set all the session variables before the last action call in the body, e.g.
`session 1`
`session.put() // session 1 updated`
`SlackPlatform.PostMessage(...) // creates session 2 (merged from session 1)`