moved DialogFlow configuration logic to a dedicated class
Extracting values from the Configuration object was polluting the DialogFlowApi class with dedicated loading methods, static values for the property keys, and static values for the default values. This is now moved in a dedicated class that summarizes all the configuration options for DialogFlow.
This commit also adds a test class for the DialogFlowConfiguration.