removed ExecutionModelUtils#getTopLevelIntents
Top-level intents are now treated as regular intents: they need a context to be triggered. This context is automatically set by the session when changing the state. It's the responsibility of the intent provider to take this information into account, and deploy it to the concrete provider if needed (this is done in the DialogFlowApi)