Set context in detectIntent instead of ContextsClient (fix #298)
ContextsClient does not allow to set multiple contexts in one request, meaning that the time needed to merge the local session to DialogFlow depends on the number of contexts to deploy. This scales very poorly (~100 ms / context).
We now set the contexts as part of the SessionsClient#detectIntent, that allows multiple context creations.