removed intent context construct from the DSL
We don't want to define multiple contexts for intents, a single context managed by Xatkit is enough.
Note that this features requires several changes at the metamodel level, for now the single context behavior is emulated with a constant context name (XATKITCONTEXT) that will be removed in a future commit.