removed event context construct from the DSL
As for intents, we don't need to define multiple contexts for events, a single context managed by Xatkit is enough.
Note that this feature requires several changes at the metamodel level, for now the single context behavior is emulated with a constant context name (XATKITCONTEXT) that will be removed in a future commit.