Removed context support in Xatkit
Intent parameters are now defined as top-level citizen in IntentDefinitions. This has several implications: the metamodel is updated to reflect this, the fluent API too (the context() construct is removed), and several internal classes needed to be updated.

IntentRecognitionProviders that rely on contexts are now setting them internally (e.g. DialogFlow works with an internal "Xatkit" context which isn't accessible to the client).

Since we don't support contexts anymore we also don't support lifespans, which means we don't need an internal data structure to retain data that has been received from a previous intent. To be concrete, I removed the RuntimeContexts class, as well as the XatkitSession. The former is the data structure that was holding these values, the latter wasn't needed anymore, the migration to SateContext was already partially implemented in the last cycle, this was a good opportunity to finish it.

Side note: the Xatkit predicates have been moved to the metamodel. This is possible because we dropped xcore in the metamodel, and we decided to manually extend the metamodel when it makes sense.