make sure events are retained when navigating context-based transitions
Context EventInstance was erased before evaluating navigable transitions after the computation of a state's body. This behavior ensured that the returned transitions do not include a transition matching the received event.  
This causes a problem if the state machines move through an AutoTransition to a state that needs to use the event (e.g. to call a reply() action). In this case the context doesn't contain an event, and the reply action crashes while looking for the channel.
Here is a simple pattern to reproduce this error (based on the ReactPlatform):
```
init
  .next()
    .when(eventIs(ReactEventProvider.ClientReady)).moveTo(greetUser);
greetUser
// The access here is fine, we come from an Event-based transition so
// the event is in the context.
  .body(context -> reactPlatform.reply(context, "...");
  .next()
    .moveTo(awaitingInput);
awaitingInput
// NPE because we come from an AutoTransition, meaning that the event
// has been removed from the context to evaluate transitions.
  .body(context -> reactPlatform.reply(context, "...");
```
Note that this issue only concerns the computation of navigable transitions that happens _after_ the computation of a state body. The computation of navigable transitions when an event is received works as expected.
*Note:* this commit is a quick fix, we should split getNavigableTransitions to handle these two use cases in a clean way.