more refactoring and add documentation
Toxicity scores are now stored in dedicated objects, and can be accessed from RecognizedIntent#getNlpData using dedicated keys:
- Perspective API: ToxicityPostProcessor#PERSPECTIVEAPI_PARAMETER_KEY
- Detoxify: ToxicityPostProcessor#DETOXIFY_PARAMETER_KEY
This commit also revert the merge of Toxicity labels into a single enumeration. A common place to store all the labels in a platform independent way is confusing and should be handled with more care.