fix VariableLoaderHelper to match Xatkit properties
The loader was using customer properties, this was not necessary. This new version reuses the properties defined in the existing *Configuration classes and aligns the test-variables file with a regular properties file.
This commit also updates the PerspectiveAPI tests to use the loader instead of ad-hoc variable loading.
Finally, this commit adds @Ignore to Detoxify tests because they cannot be integrated right now in our CI/CD pipeline.